% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_for_functions.R
\name{check_manifest_files}
\alias{check_manifest_files}
\title{Checks that the files defined in the manifest exist}
\usage{
check_manifest_files(my_manifest, file_dir)
}
\arguments{
\item{my_manifest}{A list with the contents of the manifest file.}

\item{file_dir}{Path to the directory where all the files have been extracted.}
}
\description{
Checks that the files defined in the manifest exist
}
