% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{convert_metadata_to_lists}
\alias{convert_metadata_to_lists}
\title{Fix the metadat so that it is lists}
\usage{
convert_metadata_to_lists(my_metadata)
}
\arguments{
\item{my_metadata}{A list with the information in the GoogleSheet as generated
by metadata_list_to_fsk.}
}
\description{
Fix the metadat so that it is lists
}
