% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{download_metadata_schema}
\alias{download_metadata_schema}
\title{#' Download the latest version of the MetaData Master Table as Excel}
\usage{
download_metadata_schema(out_path, sheet = NULL)
}
\arguments{
\item{out_path}{Character saying where to save the file.}

\item{sheet}{Character specifying what sheet to download. All of them by default (NULL).}
}
\value{
None
}
\description{
#' Download the latest version of the MetaData Master Table as Excel
}
