% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_fsk.R
\name{export_otherfiles}
\alias{export_otherfiles}
\title{Export other files}
\usage{
export_otherfiles(fsk_object, out_path, check = FALSE)
}
\arguments{
\item{fsk_object}{The instance of FSK2R to be exported.}

\item{out_path}{Path where the file is to be saved.}

\item{check}{Whether checks are made. TRUE by default.}
}
\description{
Export other files
}
