% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{find_packages}
\alias{find_packages}
\title{Finds where packages are stored}
\usage{
find_packages(pckgs)
}
\arguments{
\item{pckgs}{Character vector with packages names}
}
\value{
A list of packages locations. If one is not present, a character(0).
}
\description{
Finds where packages are stored
}
