% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_scope}
\alias{get_scope}
\title{Returns the scope of an FSK object}
\usage{
get_scope(fsk_obj)
}
\arguments{
\item{fsk_obj}{An object of class FSK2R}
}
\value{
A nested list with the following entries:
       \itemize{
         \item product
         \item hazard
         }
}
\description{
Returns the scope of an FSK object
}
\examples{
\donttest{
 path_example <- system.file("extdata", "ToyModelv4.fskx", package = "FSK2R")
 my_fsk <- import_fsk(path_example)
 get_scope(my_fsk)
}

}
