% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{get_session_info}
\alias{get_session_info}
\title{Extract session information}
\usage{
get_session_info()
}
\value{
A list with 3 elements: r_version, platform and pckgs.
The latter is a data.frame with two columns: package and version.
}
\description{
Extract session information
}
