% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_simulations}
\alias{get_simulations}
\title{Returns a summary of the simulations of an FSK object (NULL)}
\usage{
get_simulations(fsk_obj)
}
\arguments{
\item{fsk_obj}{An object of class FSK2R}
}
\description{
The function is not in-use. It is kept here for compatibility with older versions.
}
