% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_fsk.R
\name{import_fsk}
\alias{import_fsk}
\title{Import an FSK model into R}
\usage{
import_fsk(file_path, check = FALSE)
}
\arguments{
\item{file_path}{Path where the file is located.}

\item{check}{Whether checks are made. FALSE by default.}
}
\value{
An instance of FSK2R.
}
\description{
Importst the file in file_path and transforms it into a list of class FSK2R.
}
\examples{
\donttest{
 path_example <- system.file("extdata", "ToyModelv4.fskx", package = "FSK2R")
 my_fsk <- import_fsk(path_example)
 get_general_info(my_fsk)
}

}
