% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_nodes.R
\name{import_fsk_join}
\alias{import_fsk_join}
\title{Import of FSK with join node}
\usage{
import_fsk_join(file_path, check = TRUE)
}
\arguments{
\item{file_path}{Path where the file is located.}

\item{check}{Whether checks are made. FALSE by default.}
}
\description{
Join nodes are not yet supported by FSK2R. It just gives
an error message when called.
}
