% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{is.FSK2R}
\alias{is.FSK2R}
\title{Is it an instance of FSK2R?}
\usage{
is.FSK2R(object)
}
\arguments{
\item{object}{Object to check}
}
\value{
A logical vector
}
\description{
Is it an instance of FSK2R?
}
\examples{
\donttest{
 path_example <- system.file("extdata", "ToyModelv4.fskx", package = "FSK2R")
 my_fsk <- import_fsk(path_example)
 is.FSK2R(my_fsk)
}

}
