% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{is_fsk_with_r}
\alias{is_fsk_with_r}
\title{Does the object have an R model?}
\usage{
is_fsk_with_r(fsk_obj)
}
\arguments{
\item{fsk_obj}{An object of class FSK2R}
}
\value{
A logical vector.
}
\description{
Does the object have an R model?
}
\examples{
\donttest{
 path_example <- system.file("extdata", "ToyModelv4.fskx", package = "FSK2R")
 my_fsk <- import_fsk(path_example)
 is_fsk_with_r(my_fsk)
}

}
