% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_maps.R
\name{map_FSK_metadata}
\alias{map_FSK_metadata}
\title{Map for the contents of the metadata}
\usage{
map_FSK_metadata(type_of_model = "generic", fsk_version = "1.04")
}
\arguments{
\item{type_of_model}{Type of model, as defined in the FSK-ML documentation.
By default, 'generic'.}

\item{fsk_version}{Character stating the version of FSK-ML.}
}
\value{
A list with two components: the 'range' where each piece of information is stored and
'ws_name' with the name of the relevant sheet in the GoogleSheet template.
}
\description{
Maps the location (range) of different pieces of data within
the Excel/Google Sheets template.
It also includes the names of the sheets.
}
