% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_maps.R
\name{map_metadata_xml_template}
\alias{map_metadata_xml_template}
\title{Map between the names used in the template and the xml}
\usage{
map_metadata_xml_template()
}
\description{
Returns a map of the names used within the sheets of the
Excel/GoogleSheets template and the ones in metadata.json.
}
