% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_maps.R
\name{metadata_list_to_fsk}
\alias{metadata_list_to_fsk}
\title{From read_fsk_metadata_XX to FSK2R format}
\usage{
metadata_list_to_fsk(my_metadata, fsk_version = "1.0.5")
}
\arguments{
\item{my_metadata}{A list generated by}

\item{fsk_version}{Version of the FSK template.}
}
\description{
Converts the contents of the Excel/Google Sheets template into a list
with the format of the FSK2R object.
}
