% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{n_simuls_fsk}
\alias{n_simuls_fsk}
\title{Number of simulations in the FSK2R object}
\usage{
n_simuls_fsk(fsk_obj)
}
\arguments{
\item{fsk_obj}{An instance of FSK2R}
}
\value{
An integer vector of length one.
}
\description{
Number of simulations in the FSK2R object
}
\examples{
\donttest{
 path_example <- system.file("extdata", "ToyModelv4.fskx", package = "FSK2R")
 my_fsk <- import_fsk(path_example)
 n_simuls_fsk(my_fsk)
}

}
