% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_fsk.R
\name{read_fsk_json_metadata}
\alias{read_fsk_json_metadata}
\title{Read the metadata.json file}
\usage{
read_fsk_json_metadata(file_dir, check = FALSE, filename = "metaData.json")
}
\arguments{
\item{file_dir}{path to the file.}

\item{check}{Whether to make checks. FALSE by default.}

\item{filename}{Name of the file whith the information (meataData.json by default).}
}
\value{
A list with the contents of the metadata file.
}
\description{
Read the metadata.json file
}
