% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_fsk.R
\name{read_fsk_manifest}
\alias{read_fsk_manifest}
\title{Read the manifest of an FSK file and convert it to a data.frame}
\usage{
read_fsk_manifest(file_dir, check = FALSE, filename = "manifest.xml")
}
\arguments{
\item{file_dir}{path to the file.}

\item{check}{Whether to make checks. FALSE by default.}

\item{filename}{Name of the file whith the information (manifest.xml by default).}
}
\value{
A data.frame with the contents of the xlm file.
}
\description{
Read the manifest of an FSK file and convert it to a data.frame
}
