% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{read_fsk_metadata}
\alias{read_fsk_metadata}
\title{Reads the metadata contained in a Google Sheet}
\usage{
read_fsk_metadata(fsk_object, title, type_of_model = "generic")
}
\arguments{
\item{fsk_object}{FSK2R object where to save the metadata}

\item{title}{Character identifying the Google Sheet}

\item{type_of_model}{Character identifying the type of model.}
}
\value{
A list with the information in the GoogleSheet as generated
by metadata_list_to_fsk.
}
\description{
Reads the metadata contained in a Google Sheet
}
