% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{read_fsk_metadata_excel}
\alias{read_fsk_metadata_excel}
\title{FSK metadata from local Excel file}
\usage{
read_fsk_metadata_excel(
  fsk_object,
  path,
  type_of_model = "generic",
  fsk_version = "1.0.5"
)
}
\arguments{
\item{fsk_object}{FSK2R object where to save the data}

\item{path}{character describing the path to the file}

\item{type_of_model}{character identifying the type of model}

\item{fsk_version}{Character describing the version of FSK-ML ("1.04" by default).}
}
\value{
A list with the information in the Excel file as generated
by metadata_list_to_fsk.
}
\description{
FSK metadata from local Excel file
}
