% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_fsk.R
\name{read_fsk_model}
\alias{read_fsk_model}
\title{Read the model.sbml}
\usage{
read_fsk_model(file_dir, check = FALSE, filename = "model.sbml")
}
\arguments{
\item{file_dir}{path to the file.}

\item{check}{Whether to make checks. FALSE by default.}

\item{filename}{Name of the file whith the information (model.sbml by default).}
}
\value{
A list with the contents of the .xml file.
}
\description{
Read the model.sbml
}
