% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_fsk.R
\name{read_fsk_packages}
\alias{read_fsk_packages}
\title{Read the packages.json}
\usage{
read_fsk_packages(file_dir, check = FALSE, filename = "packages.json")
}
\arguments{
\item{file_dir}{path to the file.}

\item{check}{Whether to make checks. FALSE by default.}

\item{filename}{Name of the file whith the information (packages.json by default).}
}
\value{
A list with the contents of the JSON file.
}
\description{
Read the packages.json
}
