% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_fsk.R
\name{read_fsk_rdf_metadata}
\alias{read_fsk_rdf_metadata}
\title{Read the metadata.rdf}
\usage{
read_fsk_rdf_metadata(file_dir, check = FALSE, filename = "metadata.rdf")
}
\arguments{
\item{file_dir}{path to the file.}

\item{check}{Whether to make checks. FALSE by default.}

\item{filename}{Name of the file whith the information (metadata.rdf by default).}
}
\value{
A list with the contents of the .xml file.
}
\description{
Read the metadata.rdf
}
