% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_fsk.R
\name{read_fsk_readme}
\alias{read_fsk_readme}
\title{Read the README file}
\usage{
read_fsk_readme(file_dir, check = FALSE, filename = "README.txt")
}
\arguments{
\item{file_dir}{path to the file.}

\item{check}{Whether to make checks. FALSE by default.}

\item{filename}{Name of the file whith the information (README.txt by default).}
}
\value{
A character string with the content of the README file.
}
\description{
Read the README file
}
