% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_fsk.R
\name{read_fsk_sim}
\alias{read_fsk_sim}
\title{Read the sim.sedml file}
\usage{
read_fsk_sim(file_dir, check = FALSE, filename = "sim.sedml")
}
\arguments{
\item{file_dir}{path to the file.}

\item{check}{Whether to make checks. FALSE by default.}

\item{filename}{Name of the file whith the information (sim.sedml by default).}
}
\value{
A list with the content of the xml file.
}
\description{
Read the sim.sedml file
}
