% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_fsk.R
\name{read_other_files}
\alias{read_other_files}
\title{Read "other files"}
\usage{
read_other_files(my_tempdir, my_manifest, check = FALSE)
}
\arguments{
\item{my_tempdir}{Temporary directory to extract contents of the zyp file.}

\item{my_manifest}{A list with the information in the manifest file}

\item{check}{Whether checks are made.}
}
\description{
The R models may require further files that we can not predict. This functions just
reads all the "unrecognized" files included in the manifest and copies them to
the working directory.
}
