% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_fsk.R
\name{read_R_model}
\alias{read_R_model}
\title{Reads the R model in an FSK model}
\usage{
read_R_model(file_dir, check = FALSE, filename = "model.R")
}
\arguments{
\item{file_dir}{path to the file.}

\item{check}{Whether to make checks. FALSE by default.}

\item{filename}{Name of the file (model.R by default).}
}
\value{
A character string with the contents of the R file.
}
\description{
Reads the R model in an FSK model
}
