% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{run_simulation}
\alias{run_simulation}
\title{Run one simulation in an FSK object}
\usage{
run_simulation(fsk_object, index, run_visualization = FALSE)
}
\arguments{
\item{fsk_object}{Instance of FSK2R}

\item{index}{Index of the simulation}

\item{run_visualization}{Whether to call the visualization script. FALSE
by default.}
}
\value{
None
}
\description{
Runs the simulation corresponding to index. If defined, it also
calls any visualization script.
}
