% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{set_new_simulation}
\alias{set_new_simulation}
\title{Define a new simulation in an FSK2R object}
\usage{
set_new_simulation(fsk_object, simulation_id, parameters)
}
\arguments{
\item{fsk_object}{Instance of FSK2R}

\item{simulation_id}{A character with an id for the new simulation.}

\item{parameters}{A list whose names are the parameters to modify and their values
their values for the simulation.}
}
\value{
An instance of FSK2R with the additional simulation data.
}
\description{
Sets a new simulation using the parameters defined in simulation_pars.
The method updates all the relevant methods.
}
