% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{set_readme}
\alias{set_readme}
\title{Readme of an FSK object}
\usage{
set_readme(fsk_object, readme_text)
}
\arguments{
\item{fsk_object}{An instance of FSK2R.}

\item{readme_text}{A character vector of length 1 with the content of the README file.}
}
\value{
An instance of FSK2R.
}
\description{
Readme of an FSK object
}
\examples{
\donttest{
 path_example <- system.file("extdata", "ToyModelv4.fskx", package = "FSK2R")
 my_fsk <- import_fsk(path_example)
 set_readme(my_fsk, "This is the README.")
}

}
