% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_fsk.R
\name{update_manifest}
\alias{update_manifest}
\title{Updates the manifest file}
\usage{
update_manifest(fsk_object)
}
\arguments{
\item{fsk_object}{An instance of FSK2R.}
}
\description{
Updates the manifest file
}
