% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dataframe_to_list}
\alias{dataframe_to_list}
\title{Converts a dataframe to a list}
\usage{
dataframe_to_list(this_frame)
}
\arguments{
\item{this_frame}{data.frame to convert to a list.}
}
\description{
Stupid rjson reads data differently than idiot jsonlite, so I had
to code this shit.
}
